Move managers out from models.py in product
