+++ saleor/graphql/menu/types.py
@@ -4,7 +4,7 @@ from graphene import relay
 from ...menu import models
 from ...permission.enums import PagePermissions
 from ...permission.utils import has_one_of_permissions
-from ...product.models import ALL_PRODUCTS_PERMISSIONS
+from ...product.constants import ALL_PRODUCTS_PERMISSIONS
 from ..channel.dataloaders import ChannelBySlugLoader
 from ..channel.types import (
     ChannelContext,
